Remove fixed timings
Framerate is now the determining factor in terms of animation smoothness and update speeds for everything
* One single callback which handles rendering and updating
* Display class deals with display and has the delta time
* Sound now uses a delta time as well (it did before but now it's not a TimeSpan)
* EggHatchGUI and EvolutionGUI now use callbacks rather than a state machine
* Cleaned and made consistent usage of GL object (it's in Display now)
* Moved some stuff around namespaces and split classes
* Added constants for some things
* Vectors for colors because the in-built methods are better
* Fixed bugs with cancelling move selection when learning a move in the summary gui, and when you don't have max moves
* Cleaned up ScriptCommand so hacky workarounds aren't needed (AwaitMessage and CloseMessageBox)

+++ nukebuild/BuildWorld.cs
@@ -471,12 +471,12 @@ public sealed partial class Build
             private sealed class Stop
             {
                 private readonly int AnimTile;
-                private readonly int Time;
+                private readonly float Time;
 
                 public Stop(JToken j)
                 {
                     AnimTile = j[nameof(AnimTile)].Value<int>();
-                    Time = j[nameof(Time)].Value<int>();
+                    Time = j[nameof(Time)].Value<float>();
                 }
 
                 public void Write(EndianBinaryWriter w)
@@ -512,13 +512,13 @@ public sealed partial class Build
             }
         }
         public readonly string Tileset; // Not written
-        private readonly int Duration;
+        private readonly float Duration;
         private readonly Frame[] Frames;
 
         public TileAnimation(JToken j)
         {
             Tileset = j[nameof(Tileset)].Value<string>();
-            Duration = j[nameof(Duration)].Value<int>();
+            Duration = j[nameof(Duration)].Value<float>();
             var frames = (JArray)j[nameof(Frames)];
             int numFrames = frames.Count;
             Frames = new Frame[numFrames];
